update login functions
