+++ app/api/auth/logout/route.ts
